Pull look and feel methods out of GameRunner, clean them up.
