+++ src/games/strategy/engine/framework/startup/ui/EnginePreferences.java
@@ -37,6 +37,7 @@ import games.strategy.engine.framework.GameRunner.ProxyChoice;
 import games.strategy.engine.framework.ProcessRunnerUtil;
 import games.strategy.engine.framework.TripleAProcessRunner;
 import games.strategy.net.OpenFileUtility;
+import games.strategy.engine.framework.lookandfeel.LookAndFeel;
 import games.strategy.sound.SoundOptions;
 import games.strategy.triplea.settings.SettingsWindow;
 import games.strategy.triplea.ui.menubar.TripleAMenuBar;
